test(api): isolate MCP smoke harness tests
